Aplica ajuste de convesão de tipo para o tipo de sinais vitais ser convertido para o tipo byte na persistência do DB

+++ src/main/java/br/vet/certvet/dto/ProntuarioRequest.java
@@ -9,7 +9,7 @@ public record ProntuarioRequest (
         Clinica clinica,
         int frequenciaCardiaca,
         int frequenciaRespiratoria,
-        int temperatura,
+        byte temperatura,
         String hidratacao,
         String tpc,
         String mucosa,
